requirements.txt
black==22.10.0
click==8.1.3
cloudpickle==2.2.1
cytoolz==0.11.0
fire==0.5.0
fonttools==4.25.0
grpcio==1.49.1
ImageHash==4.3.1
Jinja2==3.1.2
line-profiler==4.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
natsort==8.3.1
pandas==1.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.1
psutil==5.9.5
Pygments==2.15.1
Pympler==1.0.1
pynvml==11.4.1
pytz==2022.2.1
PyYAML==6.0
rich==13.4.2
scalene==1.5.22.2
seaborn==0.11.2
sip==4.19.13
snakeviz==2.1.1
termcolor==2.3.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.4.0